Splay & multiple joint support (#168)
* jerry rig curl values as splay

* refactor alpha encoding to support multichar keys

* fix bug due to brackets not being alphabetic characters

* un-jerryrig curl and splay

* per joint finger curls

* aux updates

* version named pipe input

* remove last finger joint as it's never needed to be updated

* remove duplicate keys, reorder keys

* reorder splay

* fix average bone calculation

* remove unused bones from keys

* fixup inputs and comments

* glove input in named pipe

+++ src/ControllerPose.cpp
@@ -79,7 +79,7 @@ vr::DriverPose_t ControllerPose::UpdatePose() const {
       angularVelocityWorld.v[1] /= 100.0;
       angularVelocityWorld.v[2] /= 100.0;
 
-      vr::HmdQuaternion_t qAngularVelocityWorld = EulerToQuaternion(angularVelocityWorld.v[2], angularVelocityWorld.v[1], angularVelocityWorld.v[0]);
+      vr::HmdQuaternion_t qAngularVelocityWorld = EulerToQuaternion(static_cast<double>(angularVelocityWorld.v[2]), static_cast<double>(angularVelocityWorld.v[1]), static_cast<double>(angularVelocityWorld.v[0]));
 
       vr::HmdQuaternion_t qAngularVelocityObject =
           MultiplyQuaternion(MultiplyQuaternion(QuatConjugate(newPose.qRotation), qAngularVelocityWorld), newPose.qRotation);
